[feat] 전역 totalAmount를 위한 함수 위치 이동
