+++ src/constants.js
@@ -1,0 +1,1 @@
+export const INITIAL_TOTAL_AMOUNT = 0;
